third_party/blink/web_tests/platform/mac-mac10.13/virtual/threaded-composited-iframes/external/wpt/is-input-pending/tentative/same-origin-subframe.sub-expected.txt

This is a testharness.js-based test.
PASS can detect same-origin iframe events in parent
FAIL can detect same-origin iframe events in subframe assert_equals: detected discrete input expected true but got false
Harness: the test ran to completion.
